refactor: use same status pill component for main dashboard and assessments
