+++ app/components/AnalystDashboard/index.tsx
@@ -1,3 +1,3 @@
 export { default as AnalystTable } from './AnalystTable';
 export { default as AnalystRow } from './AnalystRow';
-export { default as StatusPill } from './StatusPill';
+export { default as StatusPill } from '../Analyst/StatusPill';
